sync utilities moved to util.c
